requirements.txt
ipykernel
playwright
pytest-playwright
python-dotenv